start em all

+++ ppt/start.sh
@@ -1,5 +1,9 @@
 #!/usr/bin/bash
 
+if [ $PLAYPPT -ne 0 ]; then
+ exit 1
+fi
+
 rm /tmp/.X0-lock &>/dev/null || true
 
 export DISPLAY=:0
